Add negtests and apply as macro to remove the need for asInstanceOf
